Arrumando formatação

+++ mips1.cpp
@@ -475,21 +475,21 @@ void mips1::print_cache_status(d4cache *L1) {
     std::cout << " (" <<  total_percentage << "\%" << " total)";
     std::cout << " (" << non_comp_percentage << "\%" <<" non-compulsory)"<< "\n";
 
-    std::cout << "read misses: " <<  L1->miss[D4XREAD] << "/" << (int)L1->fetch[D4XREAD];
+    std::cout << "read misses: " <<  (int)L1->miss[D4XREAD] << "/" << (int)L1->fetch[D4XREAD];
     std::cout << " (" << (double)(100*(double)L1->miss[D4XREAD]/(double)L1->fetch[D4XREAD]) << "\%)\n";
 
-    std::cout << "read non-comp misses: " <<  L1->miss[D4XREAD] - L1->comp_miss[D4XREAD] << "/" << (int)L1->fetch[D4XREAD];
+    std::cout << "read non-comp misses: " <<  (int)(L1->miss[D4XREAD] - L1->comp_miss[D4XREAD]) << "/" << (int)L1->fetch[D4XREAD];
     std::cout << " (" << (double)(100*(double)(L1->miss[D4XREAD] - L1->comp_miss[D4XREAD])/(double)L1->fetch[D4XREAD]) << "\%)\n";
 
-    std::cout << "write misses: " <<  L1->miss[D4XWRITE] << "/" << (int)L1->fetch[D4XWRITE];
+    std::cout << "write misses: " <<  (int)L1->miss[D4XWRITE] << "/" << (int)L1->fetch[D4XWRITE];
     std::cout << " (" << (double)(100*(double)L1->miss[D4XWRITE]/(double)L1->fetch[D4XWRITE]) << "\%)\n";
 
-    std::cout << "write non-comp misses: " <<  L1->miss[D4XWRITE] - L1->comp_miss[D4XWRITE] << "/" << (int)L1->fetch[D4XWRITE];
+    std::cout << "write non-comp misses: " <<  (int)(L1->miss[D4XWRITE] - L1->comp_miss[D4XWRITE]) << "/" << (int)L1->fetch[D4XWRITE];
     std::cout << " (" << (double)(100*(double)(L1->miss[D4XREAD] - L1->comp_miss[D4XWRITE])/(double)L1->fetch[D4XWRITE]) << "\%)\n";
 
-    std::cout << "compulsory misses: " << L1->comp_miss[D4XWRITE]+L1->comp_miss[D4XINSTRN]+L1->comp_miss[D4XMISC]+L1->comp_miss[D4XREAD] << "\n";
-    std::cout << "capacity misses: " <<  L1->cap_miss[D4XWRITE]+L1->cap_miss[D4XINSTRN]+L1->cap_miss[D4XMISC]+L1->cap_miss[D4XREAD] << "\n";
-    std::cout << "conflict misses: " <<  L1->conf_miss[D4XWRITE]+L1->conf_miss[D4XINSTRN]+L1->conf_miss[D4XMISC]+L1->conf_miss[D4XREAD] << "\n\n";
+    std::cout << "compulsory misses: " << (int)(L1->comp_miss[D4XWRITE]+L1->comp_miss[D4XINSTRN]+L1->comp_miss[D4XMISC]+L1->comp_miss[D4XREAD]) << "\n";
+    std::cout << "capacity misses: " <<  (int)(L1->cap_miss[D4XWRITE]+L1->cap_miss[D4XINSTRN]+L1->cap_miss[D4XMISC]+L1->cap_miss[D4XREAD]) << "\n";
+    std::cout << "conflict misses: " <<  (int)(L1->conf_miss[D4XWRITE]+L1->conf_miss[D4XINSTRN]+L1->conf_miss[D4XMISC]+L1->conf_miss[D4XREAD]) << "\n\n";
 }
 
 void mips1::detect_data_hazard() {
